Creating Bridging table
Trying to use the Primary Key ID from recipe that it started to make sure we add each ingredient to the appropriate bridging table row with the correct id. Basically trying to make sure we match up the right ingredient to the right recipe. Ran into error with SQLiteDataAccess.cs